splitting constants.py
