+++ src/extensions/progress_bar.py
@@ -1,5 +1,11 @@
-from .constants import *
-from .utils import paint_text
+from .constants import (
+    FULL_CHAR,
+    HALF_CHAR,
+    TOTAL_PROGRESS,
+    PERCENT_PROGRESS,
+    RAW_PROGRESS_BAR_LENGTH,
+)
+from ..utils import paint_text
 
 
 class ProgressBar:
