+++ src/file_operations/constants.py
@@ -1,0 +1,2 @@
+OUTPUT_FOLDER_NAME = "parser-output"
+OUTPUT_FILE_PATTERN = {"login": {}, "password": {}, "url": {}}
